feat: mv config_test copletely to internal/otelconfig
